Massive cleanup.
Vixen now has a proper CLI.  We can have pretty much arbitrary tags for
media.  The processing of data and finding the tags is separated from
the viewing and annotation process.  The annotation can be saved and
loaded from disk.  The changes to tags can also be saved from the UI.