Speed up considerably.
Turns out that when a project was scanned the media update sends a huge
number of object changed events.  We do it in one shot but that too has
problems as we cannot send too large a json string either.  This is
fixed and allows vixen to work nicely with about 25-50k files involved.