Initial support for searching media metadata.
The search facility is rather simple and brute force.  The search query
is parsed using the `whoosh` package but no indexing is done using
whoosh.  One may search for text, numeric values, numeric ranges
(including <, <=, >, >=), date ranges, and logical operations with all
of these.  Most of these are tested.  There is no UI for this yet.