Remove all direct access to project.media.
Instead we use the project.get to avoid assuming that the media
dictionary is populatd.  This will make it much easier to change the
underlying implementation and improve the ability to handle more files
easily.  One can in theory load the media lazily.  Also show the
relative path in the search browser instead of just the file name.