Use tornado v 4.x to avoid problems with jigna.
Specifically, the issue is the change to the IOLoop.instance which does
not provide a global but the current ioloop and this is a problem for us
when we make a threaded call via jigna.  We avoid this without any loss
by using a slightly older version of tornado until this is resolved in
jigna.